chore: fix PHPUnit 10 compat
