+++ src/PHPDraft/Model/Elements/Tests/RequestBodyElementTest.php
@@ -135,7 +135,7 @@ class RequestBodyElementTest extends LunrBaseTest
      *
      * @return array
      */
-    public function parseObjectProvider(): array
+    public static function parseObjectProvider(): array
     {
         $return             = [];
         $base1              = new RequestBodyElement();
